requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
aiohttp>=3.3.0,<3.4.0
websockets>=5.0,<6.0
python-Levenshtein>=0.12.0
sortedcontainers
matplotlib